Add pacstrap action
Add a pacstrap action, which is the Arch variant of debootstrap.
In other words, this allows us to bootstrap an Arch image.

Bear in mind that the tool pacstrap is a thin wrapper around the pacman
package manager. Both of those are separate projects and are available
as Debian packages, in case one wants to give them a try.

This is loosely based on the work by Arnaud Rebillout

Note: the action works only within a fakemachine instance, making it
work outside is doable but left for a later day.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
---
Some of the changes made:
 - Commit message
 - Remove all pacman.conf and mirrorlist generation
 - Use drop-in user files for ^^
 - Remove custom flags passed to pacman-key, pacstrap
 - Add Verify hook for validation
 - Sort the action
 - Run through gofmt -s -w
 - Add PreNoMachine() stage

NOTE: the funky formatting of the pacstrap_action.do header is due to
gofmt. Can drop those changes if people prefer.